# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Agent LLM Research Automation Platform API
  description: |
    Comprehensive API documentation for the Multi-Agent Research Platform.
    
    ## Architecture
    
    The platform consists of:
    - **AI Engine** (Python/FastAPI): Core agent orchestration and LLM processing
    - **Backend** (Node.js/Express): User management, research sessions, and data persistence
    - **Frontend** (React/TypeScript): User interface and visualization
    
    ## Authentication
    
    Most endpoints require JWT authentication via the `x-auth-token` header.
    
    ## Features
    
    - Multi-agent research workflows
    - Real-time streaming chat
    - Memory management
    - Web search integration
    - Export functionality (Markdown, PDF, LaTeX)
    - Token usage tracking
  version: "2.0.0"
  contact:
    name: Research Team
    email: research@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: AI Engine (Python/FastAPI)
  - url: http://localhost:5000
    description: Backend (Node.js/Express)
paths:
...
  /export/{id}/markdown:
    get:
      tags: ["Export"]
      summary: Export as Markdown
      description: Download research results as Markdown file
      security:
        - JWTAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Markdown file
          content:
            text/markdown:
              schema:
                type: string
        '404':
          description: Research not found
  /export/{id}/pdf:
    get:
      tags: ["Export"]
      summary: Export as PDF
      description: Download research results as PDF file
      security:
        - JWTAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: Research not found
  /export/{id}/latex:
    get:
      tags: ["Export"]
      summary: Export as LaTeX
      description: Download research results as LaTeX file
      security:
        - JWTAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: LaTeX file
          content:
            application/x-latex:
              schema:
                type: string
        '404':
          description: Research not found
  
  # Usage (Backend Proxy)
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in x-auth-token header
...